pyproject.toml
[project]
name = "minifram"
version = "0.1.0"
description = "Minimal agent framework for local LLM interaction"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
go = "src.foundation.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
